+++ Sources/Sentry/SentryScope.m
@@ -162,7 +162,7 @@ SentryScope ()
     [self notifyListeners];
 }
 
-- (void)setExtraValue:(id)value forKey:(NSString *)key
+- (void)setExtraValue:(id _Nullable)value forKey:(NSString *)key
 {
     @synchronized(self) {
         [self.extraDictionary setValue:value forKey:key];