+++ CHANGELOG.md
@@ -6,10 +6,13 @@ This release also enables by default the option `attackStacktrace` which include
 the stacktrace in all events, including `captureMessage` by default.
 
 Breaking Changes: 
+
 - feat: Attach stacktraces to all events by default #705 !Breaking
 
 Features and fixes: 
+
 - feat: Crash event and session in same envelope #716
+- feat: Allow nil in setExtraValue on SentryScope to remove key #703
 
 ## 6.0.0-beta.0